feat(3p/nix): Compare symbols by pointer order
A profile pass seems to indicate that comparing strings for insertion
into BtreeBindings is a performance hotspot for evaluation - the only
place where we semantically *need* to do that is lexicographicOrder,
otherwise we can compare the pointers to the intern table which is
significantly faster. Haven't profiled this yet, but will update with
metrics when I do so.

Change-Id: I80ee2b402f45e7c4c3e190e2da9024ffee9e6609