Update patch set 3
Patch Set 3:

(1 comment)

Historical note: reviewed in a stack with https://cl.tvl.fyi/c/depot/+/1246/2 .

> Patch Set 2:
> 
> So this seems to run, but it posts this:
> 
> > <input>:1:1:Previously present file "nix/buildGo/example/thing.proto" was deleted.
> 
> What does that mean? Is it keeping state between runs somewhere?

Nope, it's pulling the "previous" out of `.git#branch=canon`. Because this was originally a full stack of commits where I added `nix/buildGo` to the ignore list, it treats that as a deletion vs canon.

Patch-set: 3