feat(web/todolist): Implement a "todo-list" page generator
This invokes ripgrep & jq to construct a list of TODOs from known
users across depot sources, and dumps it into a static page that we
can serve.

The structure is relatively simple, but it might be useful. See here
for an example of what this looks like:

https: //tazj.in/blobs/todos.png
Change-Id: I1edef56606273584ab886b9e762c8ed4d210919d
Reviewed-on: https://cl.tvl.fyi/c/depot/+/1296
Tested-by: BuildkiteCI
Reviewed-by: Alyssa Ross <hi@alyssa.is>