feat: Introduce the most cursed Nix code I could think of
DO NOT SUBMIT THIS (just a warning to myself)

This passes in an optional attribute called __findFile to packages.
Packages that choose to accept this parameter suddenly find that the
meaning of '<...>' in their Nix code changes.

Specifically, an example REPL session:

    tazjin@frog /depot (canon)> nix repl .
    Welcome to Nix version 2.3.5. Type :? for help.

    Loading '.'...
    Added 22 variables.

    nix-repl> <ops/besadii>
    «derivation /nix/store/6kvn1xcy0h5nnsbggyph6h3267hs773w-besadii.drv»

A demonstration of what this looks like for e.g. Go-packages is also
included.

Change-Id: I8eea9925e34ad0c19a5935cb1848441515cfa084