Update patch set 1
Patch Set 1: Code-Review-2

This uses a feature we could otherwise probably drop completely from tvix.  Using it as a means of package access circumvents overrides, which I don't think is a good thing.  It's possible to work around that, sure, but it seems a shame to require that.  I also don't think that the per-file nature of this means that it is going to be useful for preventing looking things up in the NIX_PATH.

Patch-set: 1
Reviewer: Gerrit User 1000018 <1000018@4fdfa107-4df9-4596-8e0a-1d2bbdd96e36>
Label: Code-Review=-2