Update patch set 1
Patch Set 1:

Some of these are linked in libstore (I think this is mostly copied from the pre-cmake build system). Could we check that we're actually linking to stuff that is directly in use by the target for libstore?

Patch-set: 1
Reviewer: Gerrit User 1000001 <1000001@4fdfa107-4df9-4596-8e0a-1d2bbdd96e36>