feat(3p/nix): Add stub RPC store client
Add a stub class for wrapping a gRPC client to the new, proto-backed nix
store protocol, along with one method implemented for the sake of
demonstration. The plan is to fill all of these in and then rip out
UDSRemoteStore in favor of this.

Paired-With: Vincent Ambo <mail@tazj.in>
Paired-With: Perry Lorier <isomer@tvl.fyi>
Change-Id: Id943d4f6d75084b8498786d580e6c9f7c92c104d