feat(3p/nix): Implement AddToStore proto handler
Implement the proto handler for AddToStore, which as far as I can tell
is *actually* AddNarToStore, whereas the AddToStoreNar function looks
more like the one that is adding a file/directory to the store directly.
This is implemented by adding a new (probably soon-to-be-generalized)
Source concretion that wraps a grpc ServerReader for the stream of data
we're receiving from the client - this is less than ideal, as it's
perpetuating the source/sink thing that's going on and storing entire
nars in memory, but is at the very worst an incremental step towards a
functioning nix that we can refactor in the future.

Change-Id: I48db734e7460a47aee4a85dd5137b690980859e3