feat(3p/nix): Remove "dumping large path" warning
This has always annoyed me - 256 MiB is really not that large, and I
have never seen this run out of memory before (and even if it did, the
warning wouldn't really be helpful as things can OOM for many reasons,
not just dumping large paths). At the end of the day it just feels like
noise.

Change-Id: Iff9a51e495315fcf3e83d39fa1f2a855f5325ba3