Update patch set 1
Patch Set 1: Code-Review-1

> Patch Set 1:
> 
> I do find it useful as an indicator that something is getting included in the sources that shouldn't be (e.g. a cargo output if I did an in-tree build). Maybe the message should be changed & have a higher trigger threshold?

It’s a super helpful indicator that you are copying a source expression to the store that you really don’t want to. For example it happens frequently when you handle <nixpkgs> incorrectly, and the semantics are super tricky.

256MiB is fucking ginormous tbh, it won’t even trigger on smallish repositories where you forget to fileterSource a `.git`.

I’ve looked at the implementation multiple times in the past, and it doesn’t tell you what it’s moving that takes so much space. It would be great if we could move that to a place where it would actually tell you.

Patch-set: 1
Reviewer: Gerrit User 1000013 <1000013@4fdfa107-4df9-4596-8e0a-1d2bbdd96e36>
Label: Code-Review=-1