feat: Add rustfmt configuration for the whole depot
Most settings are left at the defaults.

There are a few more that we want, but even though a stable rustfmt is
perfectly capable of dumping them into a config file, it can not
actually use them.

I expect this will be mostly agreeable, with the exception of
`combine_control_expr`.

Change-Id: I66e3b4705e24dbba902e8847397365fcf31dd403
Reviewed-on: https://cl.tvl.fyi/c/depot/+/1555
Tested-by: BuildkiteCI
Reviewed-by: kanepyork <rikingcoding@gmail.com>
Reviewed-by: tazjin <mail@tazj.in>