fix(3p/nix): absl::StrSplit("", ...) returns {""}
When tokenizeString was changed to absl::StrSplit, there was a behavior
change because tokenizeString on an empty string returned an empty
vector - which the derivation builder (and likely a bunch of other
stuff) was depending on. In this particular case just emptying the
vector if the source string is empty works, but there may be other
places we need to fix this too.

This commit also includes some opportunistic absl::StrFormats, because I
was here anyway, but those have no semantic difference.

Change-Id: Ibf9bb602284f793fa55728481f63b838fb7a41db