fix(3p/nix): Partially revert "eliminate exposed global variable"
This reverts part of commit 976a36c2e482f416acd79a624e6d96cce2564b5b,
which was causing scanForReferences to hang indefinitely. I'm not sure
what the original intent of the commit was since there's not really much
of a commit message on it, but with this I am able to fully realise
derivations.

Change-Id: I620e892e2fe2244f066512286b086b3dd8f8f6c0