feat(ops/pipelines): Dynamically generate CI pipeline from targets
Create the pipeline by outputting a file that contains nix-build
invocations for each target's *derivation path*.

This makes it possible to run target-level granular pipelines. We're
getting somewhere.

Note: At the moment this *requires* all builds to happen on the same
machine that created the pipeline config. In practice this is true
for us, but we may want to do something about it in the future.

Change-Id: Ia6946e389dafd1d4926130bb8891446d6e17133b