feat(tvix): Support systemd socket activation
When the nix daemon starts up, first check (using sd_listen_fds) whether
we have been systemd socket-activated. If so, instead of passing the nix
daemon socket path to grpc, start a manual accept(2) loop, passing the
client file descriptors to grpc via AddInsecureChannelFromFd. There's an
open grpc issue at https://github.com/grpc/grpc/issues/19133 for
building support into grpc to do this automatically, but as of right now
this appears to be the only way to make this happen. Making this happen,
by the way, was a bit of a journey - at one point I attempted to
ServerBuilder's experimental AddExternalConnectionAcceptor API, and that
didn't work either - it appears that the final missing piece to getting
this working was explicitly fcntl(2)ing the client file descriptors to
set O_NONBLOCK before passing them into gRPC. With that set, this all
works inside of the test vm.

Fixes: b/56
Change-Id: I5d2ab2b5b02eb570249b30a9674e115c61b0ab0e