feat(tvix/daemon): accept settings overrides via grpc metadata
There's a lot of extraneous string copying in this code, but we can clean it up later.

More important to explain is the wacky levels of reflection going on. We annotate each field of the options struct with the name the setting has in the Nix code, so we can pull that out of the proto descriptors as we inspect each field from the client.

Afterward, we need some special handling of substituters and Duration-type messages, then handle the ints and bools generically.

Change-Id: I94ae2153c59c4bc4b168ca209b47b98733e7039f