refactor(tvix): Remove default buildMode arguments from libstore
These were not used anyway, and *may* have been causing a weird series
of problems leading to things sometimes aborting. Or maybe it has
nothing to do with it - the only correlation we've found is from
disassembling the output of a core dump. Regardless, this feels
stylistically like a good idea.

Change-Id: I288b50945e74ea90c118a21767fa4387c9f47f7d